Remove the hard dependency on docker
Building with lal, without docker, can lead to a simpler (internal)
implementation and can enable a quicker start for new users to lal,
who don't have complete access to the build systems and conventions
used by Cisco.

Approach:

A new struct, Environment, is introduced and replaces all cases
where the Container struct is used. Environment is an enum of either
Container (the usual behaviour) or None (representing usage without
containment). Other Environment variants could be introduced.

Testing:

The `parameterized` macro is used to set both a docker environment,
and an uncontained environment.

If testing without docker, the test cases with "alpine" are skipped.

run the full test suite with

	cargo test

or just run the non-docker containers tests with

	cargo test --no-default-features