Get started with lal easier
Remove the `configs/` and remove the need to provide a defaults file on first startup.

From a fresh checkout

	$ cargo build
	$ cargo run -- configure
	$ cargo run -- build

The README has been updated with more examples.

This is now because the manifest file can contain the environment
definitions, and lal will fallback to the user-wide environment
definitions only if not found (to preserve behaviour).

As a consequence, the environment to build code can now be configured
with the code in `.lal/manifest.json`.