Take a global lock when executing commands
This prevents a race condition when using Command::current_dir.

Although tests are run in threads, they exist in the same process.
Since a process can only exist in a single directory, all threads
will be changed to the same directory; this is a race condition.

This solution is to acquire a global lock to prevent two commands
from executing concurrently, from when the directory is changed
until the subprocess exits.