Update hyper, add tokio. Go async
This commit updates to a version of hyper that needs
support from an async runtime, and thus chooses tokio.

The need for a http client, is currently limited to the
Artifactory backend, but we may need a general way to
upload and download files (and other async work) before
we can support other backends.

Thus, the storage traits are now annotated with async_trait,
and this needs to be propagated thoughout the rest of the program
up the call stack as is typical for "coloured functions".
https://journal.stuffwithstuff.com/2015/02/01/what-color-is-your-function/

main, is now async, and so are most tests.
However, unlike main which can be annotated with the #[tokio::main] macro
tests which are parameterised cannot. The workaround is to spawn
the test inside an async block given to tokio's `block_on`.

This could be simplified in the future with a macro, but for now, I haven't
been able to find an async test macro that supports fixtures and data
parameterization.